accessment/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
atom==0.11.0
bugsnag==4.7.1
build==1.2.2.post1
CacheControl==0.14.3
cerebrium==1.40.4
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cleo==2.1.0
click==8.2.1
coloredlogs==15.0.1
crashtest==0.4.1
cryptography==45.0.3
distlib==0.3.9
dulwich==0.22.8
fastapi[standard]
fastjsonschema==2.21.1
filelock==3.18.0
findpython==0.6.3
flatbuffers==25.2.10
fsspec==2025.5.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
humanfriendly==10.0
humanize==4.12.3
idna==3.10
installer==0.7.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jeepney==0.9.0
Jinja2==3.1.6
keyring==25.6.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
more-itertools==10.7.0
mpmath==1.3.0
msgpack==1.1.0
networkx==3.5
numpy==2.2.6
onnx==1.18.0
onnxruntime==1.22.0
packaging==24.2
pbs-installer==2025.5.17
pillow==11.2.1
pkginfo==1.12.1.2
platformdirs==4.3.8
poetry==2.1.3
poetry-core==2.1.3
protobuf==6.31.1
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
pyflakes==3.3.2
Pygments==2.19.1
PyJWT==2.10.1
pyproject_hooks==1.2.0
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.13.0
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
SecretStorage==3.3.3
setuptools==80.9.0
shellingham==1.5.4
sniffio==1.3.1
starlette==0.46.2
sympy==1.14.0
tenacity==9.1.2
termcolor==2.3.0
toml==0.10.2
tomlkit==0.13.2
torch==2.7.0
torchvision==0.22.0
tqdm==4.67.1
triton==3.3.0
trove-classifiers==2025.5.9.12
typer==0.12.5
typing-inspection==0.4.1
typing_extensions==4.14.0
tzlocal==5.3.1
urllib3==2.4.0
uvicorn
virtualenv==20.31.2
WebOb==1.8.9
yaspin==3.1.0
zstandard==0.23.0
